pyproject.toml
[tool.poetry]
name = "simpet"
version = "0.1.0"
description = "The simpet project is intended to setup and launch MC simulation on a simple way."
authors = ["YerePhy <jereasap@gmail.com>, txusser <jesus.bubuchis@gmail.com>"]
readme = "README.md"
include = [
    { path = "src"},
    { path = "utils" }
]
exclude = ["**tests/**"]

[tool.poetry.dependencies]
python = "~3.9"
numpy = "1.24.1"
scipy = "1.10.1"
nibabel = "3.2.2"
pyaml = "21.10.1"
nipype = "1.8.4"
nilearn = "0.10.0"
pexpect = "4.8.0"
hydra-core  = "1.3"
pyprojroot = "0.3.0"
ruamel-yaml = "0.17.21"

[tool.poetry.group.dev.dependencies]
matplotlib = "3.6.3"
pandas = "1.5.2"
pytest = "7.1.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
